requirements.txt
# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0

# Data Science & Visualization
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Progress Bars
tqdm>=4.62.0

# Optional - for enhanced XAI features
plotly>=5.0.0  # For interactive visualizations
